refactor(core): extract createViewportSyncObserver (PLAN §3.1.i)
Pull viewport-sync out of init() into a self-contained observer:

- redetectElementUnderPointer: gated re-detection at the pointer position
- handleViewportChange: cache-invalidate + redetect + viewport-version bump
- scroll/resize/visualViewport listeners
- scheduleBoundsSync: rAF-throttled coarse re-sync
- The shouldRunInterval effect: 100ms ticker that pins overlay bounds while
  the renderer is showing something
- Disposal of both timers in onCleanup

The controller takes (grab, phase, isEnabled, isThemeEnabled,
eventListenerManager) and owns:
  - boundsRecalcIntervalId, viewportChangeFrameId, previousViewport{Width,Height}

Drops 3 now-unused imports from index.tsx (nativeCancelAnimationFrame,
nativeRequestAnimationFrame, invalidateInteractionCaches,
ZOOM_DETECTION_THRESHOLD). 39 viewport + copy-feedback specs green.

index.tsx: 3425 -> 3314 lines (-111).
